+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 
 ## Improvements
 
+- [connectivity] Add details to error message for missing "URL" properties in destinations.
 - [util] Add `sanitizeRecord` function to `cloud-sdk-logger` which replaces potentially sensitive information in a `Record<string, any>` based on a list of sensistive keys.
 
 ## Fixed Issues