+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 ## Improvements
 
 - [util] Add `sanitizeRecord` function to `cloud-sdk-logger` which replaces potentially sensitive information in a `Record<string, any>` based on a list of sensistive keys.
+- [generator] The new CLI option `licenceInPackageJson` offers the possibility to specify the license property in a generated `package.json`.     
 
 ## Fixed Issues