+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 
 - [connectivity] Add details to error message for missing "URL" properties in destinations.
 - [util] Add `sanitizeRecord` function to `cloud-sdk-logger` which replaces potentially sensitive information in a `Record<string, any>` based on a list of sensistive keys.
+- [generator] The new CLI option `licenceInPackageJson` offers the possibility to specify the license property in a generated `package.json`.     
 
 ## Fixed Issues