+++ CHANGELOG.md
@@ -22,6 +22,21 @@
 
 ## Improvements
 
+-
+
+## Fixed Issues
+
+-
+
+
+# 2.0.1
+
+Release Date: TBD<br>
+API Docs: https://sap.github.io/cloud-sdk/api/2.0.1<br>
+Blog: TBD<br>
+
+## Improvements
+
 - [connectivity] Add details to error message for missing "URL" properties in destinations.
 - [util] Add `sanitizeRecord` function to `cloud-sdk-logger` which replaces potentially sensitive information in a `Record<string, any>` based on a list of sensistive keys.
 - [generator] The new CLI option `licenceInPackageJson` offers the possibility to specify the license property in a generated `package.json`.